% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adjacentPairs}
\alias{adjacentPairs}
\title{Find the adjacent pairs in a sequence}
\usage{
adjacentPairs(start, stop, simplify = FALSE)
}
\arguments{
\item{start}{start position of the sequence}

\item{stop}{stop position of the sequence}

\item{simplify}{if true, return a matrix of the list}
}
\value{
a list with the pair sequence
}
\description{
Given a start and stop, output a paired sequence
}
\examples{
adjacentPairs(3, 10)
adjacentPairs(0, 4)
t(adjacentPairs(0, 4, TRUE))
}

