% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{ldaPlot}
\alias{ldaPlot}
\title{LDA plot}
\usage{
ldaPlot(data, groups, useColLabels = TRUE)
}
\arguments{
\item{data}{a matrix with samples on columns and dimensions on rows}

\item{groups}{the class of each column}

\item{useColLabels}{if \code{TRUE} print the column labels on the plot}
}
\value{
a \code{ggplot} object
}
\description{
Create a two-dimensional LDA plot using all variables in \code{data} to
predict \code{groups}
}

