% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{outer_join}
\alias{outer_join}
\title{Perform a full outer join}
\usage{
outer_join(left, right, by = NULL, ...)
}
\arguments{
\item{by}{the columns to do the join by.}

\item{x}{the first \code{data.frame}}

\item{y}{the second \code{data.frame}}
}
\value{
a \code{data.frame} with the full outer join
}
\description{
Peform a full outer join using dplyr \code{\link{dplyr::left_join}}. For
additional documentation, see \code{\link{dplyr::left_join}}.
}

