% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{conditionalBoxplots}
\alias{conditionalBoxplots}
\title{Conditional boxplots}
\usage{
conditionalBoxplots(df, y, x, binSize = 0.05, yQuantile = FALSE,
  xQuantile = TRUE)
}
\arguments{
\item{df}{a data.frame with variables of interest}

\item{y}{string representing the column name in df}

\item{x}{string representing the column to condition on}

\item{binSize}{how much spacing in the quantiles to include}

\item{yQuantile}{if TRUE, also plot y's quantiles}

\item{xQuantile}{if TRUE, plot x's quantiles}
}
\value{
a ggplot object
}
\description{
Plot boxplots of y conditional on the quantile of x
}

