% Generated by roxygen2 (4.0.1): do not edit by hand
\name{conditionalBoxplots}
\alias{conditionalBoxplots}
\title{Conditional boxplots}
\usage{
conditionalBoxplots(df, y, x, binSize = 0.05, yQuantile = FALSE)
}
\arguments{
\item{df}{a data.frame with variables of interest}

\item{y}{string representing the column name in df}

\item{x}{string representing the column to condition on}

\item{binSize}{how much spacing in the quantiles to include}

\item{yQuantile}{if TRUE, also plot y's quantiles}
}
\value{
a ggplot object
}
\description{
Plot boxplots of y conditional on the quantile of x
}

