% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/units.R
\name{fpkmToTpm}
\alias{fpkmToTpm}
\title{Convert a vector of FPKM to TPM}
\usage{
fpkmToTpm(x)
}
\arguments{
\item{x}{a vector of all observed FPKM values}
}
\value{
a vector of all TPM values
}
\description{
Given a vector with Fragments Per Kilobase per exon model Million, convert
the units to Transcripts Per Million (TPM)
}

