% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pcaPlot}
\alias{pcaPlot}
\title{Plot the first two principle components}
\usage{
pcaPlot(df, groups = NULL, useColLabels = TRUE, centerDf = TRUE,
  scaleDf = FALSE)
}
\arguments{
\item{df}{a data.frame with columns referring to different experimental conditions}

\item{groups}{group by which to color columns by}

\item{useColLabels}{if true, print the column names on the plot}

\item{centerDf}{center df in prcomp}

\item{scaleDf}{scale df in prcomp}
}
\value{
a ggplot object
}
\description{
Given a data.frame, plot the first two principle components and label them
}
\examples{
require(graphics)

head(USArrests)
pcaPlot(USArrests)
pcaPlot(USArrests, c('crime', 'crime', 'pop', 'crime'))
pcaPlot(USArrests, c('crime', 'crime', 'pop', 'crime'), F)
}

